# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Field Service Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{tenantId}/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads for a tenant
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /tenants/{tenantId}/leads/{leadId}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tenants/{tenantId}/leads/{leadId}/convert:
    post:
      operationId: convertLeadToJob
      tags: [leads]
      summary: Convert a lead to a job
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created job from lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /tenants/{tenantId}/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs for a tenant
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
        - in: query
          name: assignedUserId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /tenants/{tenantId}/jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tenants/{tenantId}/jobs/{jobId}/complete:
    post:
      operationId: completeJob
      tags: [completions]
      summary: Mark a job as complete with photo and signature
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteJobBody"
      responses:
        "200":
          description: Completed job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /tenants/{tenantId}/jobs/{jobId}/assign:
    patch:
      operationId: assignJob
      tags: [jobs]
      summary: Assign a job to a field staff member
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignJobBody"
      responses:
        "200":
          description: Assigned job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /tenants/{tenantId}/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats for a tenant
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /tenants/{tenantId}/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed for a tenant
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "new, contacted, quoted, converted, lost"
        priority:
          type: string
          description: "low, medium, high"
        assignedUserId:
          type: ["integer", "null"]
        estimatedValue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - customerName
        - customerEmail
        - customerPhone
        - address
        - serviceType
        - description
        - status
        - priority
        - assignedUserId
        - estimatedValue
        - notes
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        assignedUserId:
          type: ["integer", "null"]
        estimatedValue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - customerName
        - serviceType
        - status
        - priority
    UpdateLeadBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        assignedUserId:
          type: ["integer", "null"]
        estimatedValue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        leadId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
        title:
          type: string
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        address:
          type: string
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "scheduled, in_progress, completed, cancelled"
        priority:
          type: string
          description: "low, medium, high, urgent"
        assignedUserId:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        completionPhotoUrl:
          type: ["string", "null"]
        completionSignatureUrl:
          type: ["string", "null"]
        completionNotes:
          type: ["string", "null"]
        paymentStatus:
          type: string
          description: "unpaid, pending, paid"
        totalAmount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - leadId
        - quoteId
        - title
        - customerName
        - customerEmail
        - customerPhone
        - address
        - serviceType
        - description
        - status
        - priority
        - assignedUserId
        - scheduledAt
        - completedAt
        - completionPhotoUrl
        - completionSignatureUrl
        - completionNotes
        - paymentStatus
        - totalAmount
        - createdAt
        - updatedAt
    CreateJobBody:
      type: object
      properties:
        leadId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
        title:
          type: string
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        address:
          type: string
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        assignedUserId:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
      required:
        - title
        - customerName
        - address
        - serviceType
        - status
        - priority
    UpdateJobBody:
      type: object
      properties:
        title:
          type: string
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        address:
          type: string
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        assignedUserId:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
    AssignJobBody:
      type: object
      properties:
        assignedUserId:
          type: integer
      required:
        - assignedUserId
    CompleteJobBody:
      type: object
      properties:
        completionPhotoUrl:
          type: ["string", "null"]
        completionSignatureUrl:
          type: ["string", "null"]
        completionNotes:
          type: ["string", "null"]
    DashboardSummary:
      type: object
      properties:
        totalLeads:
          type: integer
        openLeads:
          type: integer
        totalJobs:
          type: integer
        activeJobs:
          type: integer
        completedJobs:
          type: integer
        totalRevenue:
          type: number
        pendingRevenue:
          type: number
        totalQuotes:
          type: integer
        pendingQuotes:
          type: integer
        conversionRate:
          type: number
      required:
        - totalLeads
        - openLeads
        - totalJobs
        - activeJobs
        - completedJobs
        - totalRevenue
        - pendingRevenue
        - totalQuotes
        - pendingQuotes
        - conversionRate
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          description: "lead_created, job_updated, quote_sent, job_completed, payment_received"
        description:
          type: string
        entityId:
          type: integer
        entityType:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - entityId
        - entityType
        - createdAt
...